requirements.txt
aws-cdk-lib
boto3
cdk-ecr-deployment
cdk-nag
constructs>=10.0.0,<11.0.0